+++ python/CHANGELOG.md
@@ -8,7 +8,8 @@
 - reset preset number of all objects when switching game modes
 - choose preset number for each puzzle
 - preset-1 is now default preset, preset-0 = silence
-- add long press functionnality to change difficulty
+- add long press functionnality to `jungle` button to change difficulty
+- add long press for shutdown and restart wifi
 
 ## [1.0.0] 2018-09-01